requirements.txt
fastapi[all]
firebase
pyrebase4
firebase_admin
mysqlclient
SQLAlchemy
pydantic